# information-srv/client/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Contain Generic Information.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Information API
  version: "1.0"
servers:
- url: /
paths:
...
  /internal/calendar/holidays:
    get:
      description: "Retrieve a list of holidays for a specified year. If no year is\
        \ provided, defaults to the current year."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_calendar_holidays_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get Holidays
      tags:
      - Calendar
  /internal/calendar/holidays/{year}:
    get:
      description: "Retrieve a list of holidays for a specified year. If no year is\
        \ provided, defaults to the current year."
      parameters:
      - description: The year to retrieve holidays
        explode: false
        in: path
        name: year
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_calendar_holidays_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get Holidays
      tags:
      - Calendar
  /internal/calendar/is-holiday/{date}:
    get:
      description: Determine if the specified date is a recognized holiday.
      parameters:
      - description: "The date to check if it is a holiday (format: YYYY-MM-DD)"
        explode: false
        in: path
        name: date
        required: true
        schema:
          format: date
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_calendar_is_holiday__date__get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Check if Date is a Holiday
      tags:
      - Calendar
  /internal/calendar/next-business-day/{date}:
    get:
      description: Retrieve the next business day following a specified date.
      parameters:
      - description: "The date to start from to find the next business day (format:\
          \ YYYY-MM-DD)"
        explode: false
        in: path
        name: date
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_calendar_next_business_day__date__get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get Next Business Day
      tags:
      - Calendar
  /internal/exchange-rate:
    get:
      description: Retrieve the exchange rate for a specified date.
      parameters:
      - description: "Start date to retrieve the exchange rate (format: YYYY-MM-DD)."
        explode: true
        in: query
        name: from
        required: true
        schema:
          format: from
          type: string
        style: form
      - description: "End date to retrieve the exchange rate (format: YYYY-MM-DD)."
        explode: true
        in: query
        name: to
        required: true
        schema:
          format: to
          type: string
        style: form
      - description: Source currency code
        explode: true
        in: query
        name: fromCur
        required: false
        schema:
          type: string
        style: form
      - description: Destination currency code
        explode: true
        in: query
        name: toCur
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_exchange_rate_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get Exchange Rates
      tags:
      - Exchange Rate
...
  /secure/calendar/holidays:
    get:
      description: "Retrieve a list of holidays for a specified year. If no year is\
        \ provided, defaults to the current year."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_calendar_holidays_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get Holidays
      tags:
      - Calendar
  /secure/calendar/holidays/{year}:
    get:
      description: "Retrieve a list of holidays for a specified year. If no year is\
        \ provided, defaults to the current year."
      parameters:
      - description: The year to retrieve holidays
        explode: false
        in: path
        name: year
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_calendar_holidays_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get Holidays
      tags:
      - Calendar
  /secure/calendar/is-holiday/{date}:
    get:
      description: Determine if the specified date is a recognized holiday.
      parameters:
      - description: "The date to check if it is a holiday (format: YYYY-MM-DD)"
        explode: false
        in: path
        name: date
        required: true
        schema:
          format: date
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_calendar_is_holiday__date__get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Check if Date is a Holiday
      tags:
      - Calendar
  /secure/calendar/next-business-day/{date}:
    get:
      description: Retrieve the next business day following a specified date.
      parameters:
      - description: "The date to start from to find the next business day (format:\
          \ YYYY-MM-DD)"
        explode: false
        in: path
        name: date
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_calendar_next_business_day__date__get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get Next Business Day
      tags:
      - Calendar
  /secure/exchange-rate:
    get:
      description: Retrieve the exchange rate for a specified date.
      parameters:
      - description: "Start date to retrieve the exchange rate (format: YYYY-MM-DD)."
        explode: true
        in: query
        name: from
        required: true
        schema:
          format: from
          type: string
        style: form
      - description: "End date to retrieve the exchange rate (format: YYYY-MM-DD)."
        explode: true
        in: query
        name: to
        required: true
        schema:
          format: to
          type: string
        style: form
      - description: Source currency code
        explode: true
        in: query
        name: fromCur
        required: false
        schema:
          type: string
        style: form
      - description: Destination currency code
        explode: true
        in: query
        name: toCur
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_exchange_rate_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get Exchange Rates
      tags:
      - Exchange Rate
...
components:
  schemas:
...
    http.exchangeRateResponse:
      example:
        date: date
        mid_rate: 5.962133916683182
        selling: 5.637376656633329
        buying_transfer: 1.4658129805029452
        buying_sight: 6.027456183070403
      properties:
        buying_sight:
          type: number
        buying_transfer:
          type: number
        date:
          type: string
        mid_rate:
          type: number
        selling:
          type: number
      type: object
    http.holidaysResponse:
      example:
        date: date
        name: name
      properties:
        date:
          type: string
        name:
          type: string
      type: object
    pkg.Response:
      example:
        detail: detail
        title: title
        status: 6
      properties:
        detail:
          type: string
        status:
          type: integer
        title:
          type: string
      type: object
...
    _internal_calendar_holidays_get_200_response:
      allOf:
      - $ref: '#/components/schemas/pkg.Response'
      - properties:
          data:
            items:
              $ref: '#/components/schemas/http.holidaysResponse'
            type: array
        type: object
      example:
        data:
        - date: date
          name: name
        - date: date
          name: name
        detail: detail
        title: title
        status: 0
    _internal_calendar_is_holiday__date__get_200_response:
      allOf:
      - $ref: '#/components/schemas/pkg.Response'
      - properties:
          data:
            type: boolean
        type: object
      example:
        data: true
        detail: detail
        title: title
        status: 0
    _internal_calendar_next_business_day__date__get_200_response:
      allOf:
      - $ref: '#/components/schemas/pkg.Response'
      - properties:
          data:
            type: string
        type: object
      example:
        data: data
        detail: detail
        title: title
        status: 0
    _internal_exchange_rate_get_200_response:
      allOf:
      - $ref: '#/components/schemas/pkg.Response'
      - properties:
          data:
            items:
              $ref: '#/components/schemas/http.exchangeRateResponse'
            type: array
        type: object
      example:
        data:
        - date: date
          mid_rate: 5.962133916683182
          selling: 5.637376656633329
          buying_transfer: 1.4658129805029452
          buying_sight: 6.027456183070403
        - date: date
          mid_rate: 5.962133916683182
          selling: 5.637376656633329
          buying_transfer: 1.4658129805029452
          buying_sight: 6.027456183070403
        detail: detail
        title: title
        status: 0
...